refactor(ui): flatten src to api, pages, components, helpers
- Merge types into api/types.ts; ui/* into helpers/runApi.ts
- Merge lib/* into helpers/flagModel.ts; constants + router under helpers/
- pages/flagPage.ts, pages/flagsList.ts (was flag/, flags/)
- Remove types/, ui/, lib/, flag/, flags/, router/ top-level dirs
- AGENTS.md layout updated
