+++ browser/flagr-ui/src/helpers/constants.ts
