--- browser/flagr-ui/src/lib/flagGuards.ts
@@ -1,7 +1,0 @@
-import type { FlagView, Variant } from '@/types'
-
-export function variantUsedInDistribution(flag: FlagView, variantId: number): boolean {
-  return (flag.segments ?? []).some((s) =>
-    (s.distributions ?? []).some((d) => d.variantID === variantId),
-  )
-}
