--- browser/flagr-ui/src/types/index.ts
@@ -1,3 +1,0 @@
-export * from './flag'
-export * from './evaluation'
-export * from './operators'
