--- browser/flagr-ui/src/types/operators.ts
@@ -1,13 +1,0 @@
-export type OperatorValue =
-  | 'EQ'
-  | 'NEQ'
-  | 'LT'
-  | 'LTE'
-  | 'GT'
-  | 'GTE'
-  | 'EREG'
-  | 'NEREG'
-  | 'IN'
-  | 'NOTIN'
-  | 'CONTAINS'
-  | 'NOTCONTAINS'
